Rename renderer files to use same naming
