+++ src/components/app-content-renderer/config-try-learn-renderer.js
@@ -1,5 +1,6 @@
-import { Grid, GridItem } from '@patternfly/react-core';
 import React from 'react';
+import { Grid, GridItem } from '@patternfly/react-core';
+
 import ConfigTryLearnTile from './config-try-learn-tile';
 
 const configTryLearnRenderer = (sections) => (
