+++ src/components/app-content-renderer/styles/Panels.scss
@@ -1,6 +1,6 @@
 /* stylelint-disable no-descending-specificity */
 .ins-c-page__landing-content {
-  .level-wrapper { 
+  .level-wrapper {
     flex-basis: 1160px; // set max width for desktop at 1450px total
     padding: var(--pf-global--spacer--xl) var(--pf-global--spacer--lg) 0 var(--pf-global--spacer--lg);
   }
